requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
streamlit==1.55.0
openai==2.29.0
anthropic==0.85.0
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
matplotlib==3.10.8
